Extracting Core Claims from Scientific Articles
This project attempts to extract a single core claim from a scientific
article. With the help of AIDA rules, these claims are extracted and
rewritten when need be. First, the core sentence of the scientific
article is extracted. Second, that sentence is checked on whether it
complies with the rules of an AIDA sentence. Third, when it does not
comply, the sentence is rewritten to fulfill the requirements. One then
(hopefully) ends up with the core claim of an article.